allow undefined in JsonCompatible for optional property interoperability
