+++ libraries/ts-json-base/src/packlets/json/common.ts
@@ -66,13 +66,46 @@ type IsUnknown<T> = unknown extends T ? ([T] extends [unknown] ? true : false) :
 
 /**
  * A constrained type that is compatible with JSON serialization.
+ *
+ * This type transforms input types to ensure they can be safely serialized to JSON:
+ * - JSON primitives (string, number, boolean, null) are preserved as-is
+ * - `undefined` is allowed for TypeScript compatibility with optional properties
+ * - Objects are recursively transformed with all properties made JSON-compatible
+ * - Arrays are transformed to contain only JSON-compatible elements
+ * - Functions are transformed to error types
+ * - Other non-JSON types are transformed to error types
+ *
+ * Note: While `undefined` is technically not JSON-serializable, it's allowed here
+ * to support TypeScript's optional property patterns. Use `sanitizeJsonObject`
+ * to remove undefined properties before actual JSON serialization.
+ *
  * @param T - The type to be constrained
  * @returns A constrained type that is compatible with JSON serialization.
+ *
+ * @example
+ * ```typescript
+ * interface IUser {
+ *   name: string;
+ *   email?: string; // Optional property can be undefined
+ * }
+ *
+ * type UserCompatible = JsonCompatible<IUser>; // Allows undefined for email
+ *
+ * const user: UserCompatible = {
+ *   name: "John",
+ *   email: undefined // This works
+ * };
+ *
+ * // Before JSON serialization, sanitize to remove undefined:
+ * const sanitized = sanitizeJsonObject(user); // Removes undefined properties
+ * JSON.stringify(sanitized.value); // Safe to serialize
+ * ```
+ *
  * @public
  */
 export type JsonCompatible<T> = IsUnknown<T> extends true
   ? JsonValue
-  : T extends JsonPrimitive
+  : T extends JsonPrimitive | undefined
   ? T
   : T extends Array<unknown>
   ? JsonCompatibleArray<T[number]>
